Use a symlink rather than a bindmount!
This works just as well, and is 'cheaper' long term. If we change
the underlying mount, the mounted paths change as well. If we
unmount it, the paths all become invalid quickly. This allows
for quick operations on NFS client mounts. If these were bind mounted
then each pod should be treated like it has a full NFS mount...

Thanks to @chasemp & @madhuvishy for the NFS wisdom!